+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/nathanab/beacon/compare/master...HEAD)
 - Update admin page with sortable and searchable date table
 
+## [v2.4.2](https://github.com/nathanab/beacon/compare/v2.4.1...v2.4.2)
+- Fix focus date label to be date name instead of date ID
+
 ## [v2.4.1](https://github.com/nathanab/beacon/compare/v2.4.0...v2.4.1)
 - Fix focus date hook causing page reload